refactor(core): move the write funnels out of core/common.py
write_parquet_with_metadata and write_geoparquet_table, plus the edges and
KV-metadata helpers they own, become core/parquet_write.py -- a sibling of the
write facade, not part of it. Pure relocation; every name is still importable
from core.common.
